+++ src/main/java/org/folio/search/service/reindex/jdbc/InstanceChildResourceRepository.java
@@ -1,0 +1,12 @@
+package org.folio.search.service.reindex.jdbc;
+
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
+
+public interface InstanceChildResourceRepository {
+
+  void deleteByInstanceIds(List<String> instanceIds);
+
+  void saveAll(Set<Map<String, Object>> entities, List<Map<String, Object>> entityRelations);
+}