+++ src/test/java/org/folio/search/service/reindex/jdbc/ContributorRepositoryIT.java
@@ -3,9 +3,7 @@ package org.folio.search.service.reindex.jdbc;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.tuple;
 import static org.folio.search.utils.SearchUtils.AUTHORITY_ID_FIELD;
-import static org.folio.search.utils.SearchUtils.SUBJECT_SOURCE_ID_FIELD;
-import static org.folio.search.utils.SearchUtils.SUBJECT_TYPE_ID_FIELD;
-import static org.folio.search.utils.SearchUtils.SUBJECT_VALUE_FIELD;
+import static org.folio.search.utils.SearchUtils.CONTRIBUTOR_TYPE_FIELD;
 import static org.folio.search.utils.TestConstants.TENANT_ID;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyCollection;
@@ -19,10 +17,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
-import org.assertj.core.api.Condition;
 import org.folio.search.configuration.properties.ReindexConfigurationProperties;
-import org.folio.search.model.reindex.UploadRangeEntity;
-import org.folio.search.model.types.ReindexEntityType;
 import org.folio.search.utils.JsonConverter;
 import org.folio.spring.FolioExecutionContext;
 import org.folio.spring.FolioModuleMetadata;
@@ -44,18 +39,18 @@ import org.springframework.test.context.jdbc.Sql;
 @EnablePostgres
 @AutoConfigureJson
 @AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
-class SubjectRepositoryIT {
+class ContributorRepositoryIT {
 
   private @SpyBean JdbcTemplate jdbcTemplate;
   private @MockBean FolioExecutionContext context;
-  private SubjectRepository repository;
+  private ContributorRepository repository;
   private ReindexConfigurationProperties properties;
 
   @BeforeEach
   void setUp() {
     properties = new ReindexConfigurationProperties();
     var jsonConverter = new JsonConverter(new ObjectMapper());
-    repository = spy(new SubjectRepository(jdbcTemplate, jsonConverter, context, properties));
+    repository = spy(new ContributorRepository(jdbcTemplate, jsonConverter, context, properties));
     when(context.getFolioModuleMetadata()).thenReturn(new FolioModuleMetadata() {
       @Override
       public String getModuleName() {
@@ -71,61 +66,8 @@ class SubjectRepositoryIT {
   }
 
   @Test
-  void getUploadRanges_returnEmptyList_whenNoUploadRangesAndNotPopulate() {
-    // act
-    var ranges = repository.getUploadRanges(false);
-
-    // assert
-    assertThat(ranges).isEmpty();
-  }
-
-  @Test
-  @Sql("/sql/populate-subjects.sql")
-  void getUploadRanges_returnList_whenNoUploadRangesAndNotPopulate() {
-    // arrange
-    properties.setUploadRangeLevel(1);
-
-    // act
-    var ranges = repository.getUploadRanges(true);
-
-    // assert
-    assertThat(ranges)
-      .hasSize(15)
-      .are(new Condition<>(range -> range.getEntityType() == ReindexEntityType.SUBJECT, "subject range"))
-      .extracting(UploadRangeEntity::getLower, UploadRangeEntity::getUpper)
-      .startsWith(tuple("0", "1"))
-      .contains(tuple("a", "b"))
-      .endsWith(tuple("e", "f"));
-  }
-
-  @Test
-  @Sql("/sql/populate-subjects.sql")
-  void fetchBy_returnListOfMaps() {
-    // act
-    var ranges = repository.fetchByIdRange("20", "21");
-
-    // assert
-    assertThat(ranges)
-      .hasSize(2)
-      .allMatch(map -> map.keySet().containsAll(List.of("id", "value", "authorityId", "instances")))
-      .extracting("value", "authorityId", "sourceId", "typeId")
-      .containsExactlyInAnyOrder(
-        tuple(
-          "Alternative History",
-          null,
-          "a5a0b02e-c868-4074-ab01-348a4e87fd9f",
-          "b36c89f9-79fe-4a0a-bc13-02d95e032c08"),
-        tuple(
-          "History",
-          "79144653-7a98-4dfb-aa6a-13ad49e80952",
-          null,
-          null)
-      );
-  }
-
-  @Test
-  @Sql("/sql/populate-subjects.sql")
-  void deleteByInstanceIds_oneSubjectRemovedAndOneInstanceCounterDecremented() {
+  @Sql("/sql/populate-contributors.sql")
+  void deleteByInstanceIds_oneContributorRemovedAndOneInstanceCounterDecremented() {
     var instanceIds = List.of("a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11", "b3bae8a9-cfb1-4afe-83d5-2cdae4580e07");
 
     // act
@@ -134,21 +76,23 @@ class SubjectRepositoryIT {
     // assert
     var ranges = repository.fetchByIdRange("0", "50");
     assertThat(ranges)
-      .hasSize(16)
-      .extracting("value", "instances")
+      .hasSize(1)
+      .extracting("name", "instances")
       .contains(
         tuple("Sci-Fi", List.of(
-          Map.of("count", 1, "shared", true, "tenantId", "consortium"),
-            Map.of("count", 1, "shared", false, "tenantId", "member_tenant"))));
+          Map.of("count", 1, "shared", true, "tenantId", "consortium",
+            "typeId", List.of("aab8fff4-49c6-4578-979e-439b6ba3600b")),
+            Map.of("count", 1, "shared", false, "tenantId", "member_tenant",
+              "typeId", List.of("9ec55e4f-6a76-427c-b47b-197046f44a53")))));
   }
 
   @Test
   void saveAll() {
-    var entities = Set.of(subjectEntity("1"), subjectEntity("2"));
+    var entities = Set.of(contributorEntity("1"), contributorEntity("2"));
     var entityRelations = List.of(
-      subjectRelation("a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11", "1"),
-      subjectRelation("b3bae8a9-cfb1-4afe-83d5-2cdae4580e07", "2"),
-      subjectRelation("9ec55e4f-6a76-427c-b47b-197046f44a54", "2"));
+      contributorRelation("a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11", "1"),
+      contributorRelation("b3bae8a9-cfb1-4afe-83d5-2cdae4580e07", "2"),
+      contributorRelation("9ec55e4f-6a76-427c-b47b-197046f44a54", "2"));
 
     repository.saveAll(entities, entityRelations);
 
@@ -156,10 +100,12 @@ class SubjectRepositoryIT {
     var ranges = repository.fetchByIdRange("0", "50");
     assertThat(ranges)
       .hasSize(2)
-      .extracting("value", "instances")
+      .extracting("name", "instances")
       .contains(
-        tuple("value1", List.of(Map.of("count", 1, "shared", false, "tenantId", TENANT_ID))),
-        tuple("value2", List.of(Map.of("count", 2, "shared", false, "tenantId", TENANT_ID))));
+        tuple("name1", List.of(Map.of("count", 1, "shared", false, "tenantId", TENANT_ID,
+          "typeId", List.of("b7df83a1-8b15-46c1-9a4c-9d2dbb3cf4d5")))),
+        tuple("name2", List.of(Map.of("count", 2, "shared", false, "tenantId", TENANT_ID,
+          "typeId", List.of("b7df83a1-8b15-46c1-9a4c-9d2dbb3cf4d5")))));
   }
 
   @Test
@@ -170,20 +116,20 @@ class SubjectRepositoryIT {
     saveAll();
   }
 
-  private Map<String, Object> subjectEntity(String id) {
+  private Map<String, Object> contributorEntity(String id) {
     return Map.of(
       "id", id,
-      SUBJECT_VALUE_FIELD, "value" + id,
-      AUTHORITY_ID_FIELD, "b7df83a1-8b15-46c1-9a4c-9d2dbb3cf4d6",
-      SUBJECT_SOURCE_ID_FIELD, "b7df83a1-8b15-46c1-9a4c-9d2dbb3cf4d5",
-      SUBJECT_TYPE_ID_FIELD, "b7df83a1-8b15-46c1-9a4c-9d2dbb3cf4d4"
+      "name", "name" + id,
+      "nameTypeId", "b7df83a1-8b15-46c1-9a4c-9d2dbb3cf4d6",
+      AUTHORITY_ID_FIELD, "b7df83a1-8b15-46c1-9a4c-9d2dbb3cf4d5"
     );
   }
 
-  private Map<String, Object> subjectRelation(String instanceId, String subjectId) {
+  private Map<String, Object> contributorRelation(String instanceId, String contributorId) {
     return Map.of(
       "instanceId", instanceId,
-      "subjectId", subjectId,
+      "contributorId", contributorId,
+      CONTRIBUTOR_TYPE_FIELD, "b7df83a1-8b15-46c1-9a4c-9d2dbb3cf4d5",
       "tenantId", TENANT_ID,
       "shared", false
     );